@@ -1,31 +1,30 @@
-name: client_flutter
-description: Flutter client for GrandStack
+name: poimen
+description: Flutter client for Poimen App
 
 publish_to: "none" # Remove this line if you wish to publish to pub.dev
 
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  cupertino_icons: ^0.1.3
+  cupertino_icons: ^1.0.5
 
-  graphql_flutter: ^3.1.0
-  json_annotation: ^3.0.1
-  smooth_star_rating: ^1.1.1
+  graphql_flutter: ^5.1.0
+  json_annotation: ^4.0.1
   conventional_commit: ^0.4.3+1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  build_runner: ^1.10.1
-  flutter_launcher_icons: ^0.7.3
-  json_serializable: ^3.4.0
+  build_runner: ^2.1.2
+  flutter_launcher_icons: ^0.9.3
+  json_serializable: ^4.1.4
 
 # The following section is specific to Flutter.
 flutter:
